requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
crispy-bootstrap5==0.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-bootstrap==0.2.4
django-bootstrap-modal-forms==2.2.0
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-filter==21.1
django-import-export==2.8.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
